Início do módulo de artigos

+++ app/Http/Controllers/Archive/ArchiveExtensionController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Archive;
 
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Models\Archive\Archive;
 use App\Models\Archive\ArchiveExtension;
 
 class ArchiveExtensionController extends Controller
@@ -39,18 +40,27 @@ class ArchiveExtensionController extends Controller
     public function delete($id)
     {
         $extensao = ArchiveExtension::find($id);
-        
-        if($extensao->delete()){
+        $verifica_registros = Archive::where('extensao_arquivo_id', $id)->get();
+
+        if(count($verifica_registros) > 0){
             return response()->json([
-                'message' => "Extensão {$extensao->extensao} excluída com sucesso!",
-                'deleted' => true
-            ]);
-        }else{
-            return response()->json([
-                'message' => "Falha ao extensão extensão!",
+                'message' => "Existem arquivos cadastrados com essa extensão, favor excluí-los!",
                 'deleted' => false
             ]);
+        }else{
+            if($extensao->delete()){
+                return response()->json([
+                    'message' => "Extensão {$extensao->extensao} excluída com sucesso!",
+                    'deleted' => true
+                ]);
+            }else{
+                return response()->json([
+                    'message' => "Falha ao extensão extensão!",
+                    'deleted' => false
+                ]);
+            }
         }
+                
     }
 
 }
